Split openrouter view by key type
Extract standard-key and management-key rendering into dedicated files.
openrouter_view.go now owns the entry point and routes to the matching
view based on u.Key.IsManagementKey. Each view owns its own limit bar
(compact vs full) so the dispatcher only handles loading/error/header.
